chore: rename MarkerCollection module name and test
