+++ junifer/pipeline/__init__.pyi
@@ -10,4 +10,4 @@ from . import registry
 from .pipeline_step_mixin import PipelineStepMixin
 from .update_meta_mixin import UpdateMetaMixin
 from .workdir_manager import WorkDirManager
-from .collection import MarkerCollection
+from .marker_collection import MarkerCollection
